move contact action modal to seperate component
