+++ src/pages/GoalsPage/ContactsPage.tsx
@@ -1,69 +1,12 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate, useSearchParams } from "react-router-dom";
-import { deleteContact, getAllContacts } from "@src/api/ContactsAPI";
+import { getAllContacts } from "@src/api/ContactsAPI";
 import Contacts from "@src/helpers/Contacts";
 import AppLayout from "@src/layouts/AppLayout";
 import ContactItem from "@src/models/ContactItem";
-import ActionDiv from "@components/GoalsComponents/MyGoalActions/ActionDiv";
-import ZModal from "@src/common/ZModal";
-import { useSetRecoilState } from "recoil";
-import { displayToast, lastAction } from "@src/store";
-import { useTranslation } from "react-i18next";
-import { useQuery, useQueryClient } from "react-query";
+import { useQuery } from "react-query";
 import { PageTitle } from "@src/constants/pageTitle";
-import EditContactModal from "./components/modals/EditContactModal";
-
-const Actions = ({ contact }: { contact: ContactItem }) => {
-  const { t } = useTranslation();
-  const setLastAction = useSetRecoilState(lastAction);
-  const setShowToast = useSetRecoilState(displayToast);
-  const [showEditModal, setShowEditModal] = useState(false);
-  const queryClient = useQueryClient();
-
-  const handleDeleteContact = async (e: React.MouseEvent) => {
-    e.stopPropagation();
-    const res = await deleteContact(contact);
-    setLastAction("partnersRevalidate");
-    setShowToast({
-      open: true,
-      message: res.message,
-      extra: "",
-    });
-    queryClient.invalidateQueries({ queryKey: ["partners"] });
-    window.history.back();
-  };
-
-  return (
-    <>
-      {showEditModal && (
-        <EditContactModal
-          contact={contact}
-          onClose={() => {
-            setShowEditModal(false);
-            queryClient.invalidateQueries({ queryKey: ["partners"] });
-            window.history.back();
-          }}
-        />
-      )}
-      <ZModal open width={400} type="interactables-modal">
-        <div style={{ textAlign: "left" }} className="header-title">
-          <p className="ordinary-element" id="title-field">
-            {contact.name}
-          </p>
-        </div>
-        <div style={{ display: "grid", gridTemplateColumns: "1fr 1fr" }}>
-          <button type="button" className="goal-action-archive shareOptions-btn" onClick={handleDeleteContact}>
-            <ActionDiv label={t("Delete")} icon="Delete" />
-          </button>
-
-          <button type="button" className="goal-action-archive shareOptions-btn" onClick={() => setShowEditModal(true)}>
-            <ActionDiv label={t("Edit")} icon="Edit" />
-          </button>
-        </div>
-      </ZModal>
-    </>
-  );
-};
+import ContactActionModal from "./components/modals/ContactActionModal";
 
 const ContactsPage = () => {
   const [searchParams] = useSearchParams();
@@ -84,7 +27,7 @@ const ContactsPage = () => {
 
   return (
     <AppLayout title={PageTitle.Contacts}>
-      {showOptions && <Actions contact={selectedContact} />}
+      {showOptions && <ContactActionModal contact={selectedContact} />}
       <div className="myGoals-container">
         <div className="my-goals-content">
           <div className="d-flex f-col">
